+++ src/block-types/header/save.js
@@ -1,6 +1,7 @@
 import { useBlockProps, RichText } from "@wordpress/block-editor";
 import typography from "./../../helpers/typography";
 import textAlignment from "./../../helpers/text-alignment";
+import blockAlignment from "./../../helpers/block-alignment";
 import colors from "./../../helpers/colors";
 import size from "./../../helpers/size";
 import spacing from "./../../helpers/spacing";
@@ -12,11 +13,11 @@ const save = (props) => {
 	} = props;
 
 	const blockProps = useBlockProps.save({
-		className: `${textAlignment.cssClasses(props)}${colors.cssClasses(
-			props
-		)}${typography.cssClasses(props)}${size.cssClasses(
-			props
-		)}${spacing.cssClasses(props)}`.trimStart(),
+		className: `${textAlignment.cssClasses(props)}${blockAlignment.cssClasses(
+			props,
+		)}${colors.cssClasses(props)}${typography.cssClasses(
+			props,
+		)}${size.cssClasses(props)}${spacing.cssClasses(props)}`.trimStart(),
 		style: {
 			...typography.styles(props),
 			...typography.cssVars(props, "header"),