+++ src/block-types/section/save.js
@@ -1,6 +1,7 @@
 import { useBlockProps, InnerBlocks } from "@wordpress/block-editor";
 import colors from "./../../helpers/colors";
 import border from "./../../helpers/border";
+import blockAlignment from "./../../helpers/block-alignment";
 import size from "./../../helpers/size";
 import spacing from "./../../helpers/spacing";
 import htmlAttrs from "./../../helpers/html-attrs";
@@ -12,8 +13,10 @@ const save = (props) => {
 
 	const blockProps = useBlockProps.save({
 		className: `${colors.cssClasses(props)}${border.cssClasses(
-			props
-		)}${size.cssClasses(props)}${spacing.cssClasses(props)}`.trimStart(),
+			props,
+		)}${blockAlignment.cssClasses(props)}${size.cssClasses(
+			props,
+		)}${spacing.cssClasses(props)}`.trimStart(),
 		style: {
 			...colors.cssVars(props, "section"),
 			...border.cssVars(props, "section"),